# src/main/java/com/example/bankcards/dto/CardDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.Status;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.RequiredArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDate;

@Schema(description = "Card DTO")
@Data
@RequiredArgsConstructor
@AllArgsConstructor
public class CardDto {
    @Schema(description = "Card ID", example = "12")
    private Long id;

    @Schema(description = "Masked card number", example = "**** **** **** 1077")
    private String cardNumber;

    @Schema(description = "Expired date", example = "2030-07-03")
    private LocalDate expiredDate;

    @Schema(description = "Card balance", example = "12000")
    private BigDecimal balance;

    @Schema(description = "Status", example = "ACTIVE")
    private Status status;

    @Schema(description = "User ID", example = "51")
    private Long userId;
}